This update will add:
1. Small changes(

+++ LearningJDBC/src/main/java/by/moiseenko/jdbc/ProductDao.java
@@ -19,7 +19,7 @@ public interface ProductDao {
 
   Product findProduct(long id);
 
-  long updateProduct(Product product);
+  int updateProduct(Product product);
 
   void deleteProduct(long id);
 }
